feat: add concrete errors to public API
This commit adds four concrete errors to the public API and ensures that
all returned errors are either instances of those concrete types OR are
wrapped by more descriptive errors.

The four error types are:

1. ClientError: for all errors that are the result of the client making
   semantic error in their request

2. ServerError: for all the (uncommon) cases where the server returns
   data that is somehow and unexpectedly invalid.

3. APIError: for any errors that occurs when interacting with the SQL
   Admin API. These errors wrap the underlying
   google.golang.org/api/googleapi.Error types.

4. DialError: for any error that occurs when attempting to connect to a
   particular SQL instance.

By providing concrete errors or wrapping concrete errors, we allow our
clients to uses the errors.As API to possibly react differently to
errors coming out of the dialer.