fix: move MySQL liveness check into driver code (#417)
The MySQL driver relies on the SyscallConn interface to perform a
liveness check of connections to MySQL. Previously, the dialer included
an implementation of this interface. A recent version of pgx, however,
also now relies on this interface. Because the MySQL-centric
implementation was never meant to be used for reads and writes, when pgx
would call it, the connection would hang.

This commit moves the MySQL-specific liveness check into the MySQL
driver wrapper to ensure that support for the check remains, but that it
does not interfere with pgx.