fix: ensure cert refresh recovers from sleep
When a computer wakes from sleep after more than ~1 hour and a client
tries to connect, the internal certificate cache will have an expired
certificate. In the past, we erroneously believed that the TLS handshake
would fail if the client certificate expired. However, with TLS 1.3, the
client is the last to send a handshake message, and so the server is
unable to report any errors until the first client read. See
https://go.dev/doc/go1.12#tls_1_3.

As a result, clients who tried to connect after sleep, would see bad
certificate errors which would resolve only once the remaining refresh
duration expired (sleep time + remaining duration).

This commit adds code to check if the client certificate is expired and
immediately triggers a block refresh if it is. This means that in the
case where a cached certificate has expired, it will be discarded and a
new one will be retrieved before continuing with the connection.

Related to https://github.com/GoogleCloudPlatform/cloud-sql-proxy/issues/1788