refactor: clean up connection info cache interface (#768)
This PR shrinks the interface for the connection info cache. A smaller interface makes it easier to provide alternate implementations. Separately, the ConnectionInfo provides a super set of the InstanceEngineVersion data, making InstanceEngineVersion a duplicate method in effect. Also, connection count tracking shouldn't be done at the cache layer and instead should be done at the dialer layer. As a result, this PR removes two methods from the interface for a tighter and clean abstraction.