feat: invalidate cache on failed IP lookup (#812)
In #768, we removed an implicit behavior of the Go Connector. If a dial
attempt requests a non-existent IP type (e.g., client asks for public IP
on a private IP only instance), the Connector would invalidate the
cache. But with the cleanup PR, we removed that implicit behavior.

In some cases, it might be useful to have this behavior. For example, if
a caller starts the Go Connector and tries to connect to a public IP and
then later configures public IP, there is no need for a restart.

We made that change in the AlloyDB Go Connector mostly because some
internal tests depend on that behavior. See
GoogleCloudPlatform/alloydb-go-connector#555.

Fixes #780