refactor: improve naming and key handling in API client
This commit uses a more descriptive name for the code that refreshes
connection info. Instead of "refresher," this commit renames the object
to "adminAPIClient" to match what we're doing in other connectors.

In addition, the RSA key is no longer stored on the cache but instead
passed to the only code that needs the key: the API client.