test: Adds test to simulate the way that the Auth Proxy calls the dialer
Added test TestDialerClosesOldConnectionsAfterDnsChange which will demonstrate
that the dialer behaves correctly when it conencts using a DNS name, and then 
the value of the DNS record changes.
The dialer should
- close existing connections, 
- open new connections to the updated instance. 
- new connections should stay open after the DNS loop completes its first cycle. 

This also makes the mock resolver more realistic. If it is called with a context that
is done, the resolver will return an error.