fix: Create a new context for the domain name check loop.
When a monitoredCache runs a goroutine to check for changes in the domain
name, it needs its own context. The proxy (and other library users) call
dialer.Dial(ctx...), they pass in a context with a 30 second timeout. 

The background process that checks for domains should not use this, it should
create its own context with no timeout. 

This fixes a customer-discovered bug in the proxy.