fix: Use a new context for the domain name check loop. (#1007)
This fixes a bug in the go connector that was exposed by the Auth Proxy. If
the Auth Proxy is configured with a domain name, and then the DNS record is
changed to point to a new instance. Connections to the new instance will be
terminated every 30 seconds.

This was caused by a bug in how the MonitoredCache set up the background
Context for for the goroutine that checks for changes in the DNS record.

**test: Adds test to simulate the way that the Auth Proxy calls the dialer**

Added test TestDialerClosesOldConnectionsAfterDnsChange which will demonstrate
that the dialer behaves correctly when it conencts using a DNS name, and then 
the value of the DNS record changes.
The dialer should
- close existing connections, 
- open new connections to the updated instance. 
- new connections should stay open after the DNS loop completes its first cycle. 

This also makes the mock resolver more realistic. If it is called with a context that
is done, the resolver will return an error.

**fix: Create a new context for the domain name check loop. **

When a monitoredCache runs a goroutine to check for changes in the domain
name, it needs its own context. The proxy (and other library users) call
dialer.Dial(ctx...), they pass in a context with a 30 second timeout. 

The background process that checks for domains should not use this, it should
create its own context with no timeout.